+++ CHANGES.rst
@@ -16,6 +16,7 @@ Unreleased
   * Add :attr:`Entry.feed_resolved_title` and :attr:`Feed.resolved_title` properties.
   * The ``feed`` search column now indexes :attr:`Entry.feed_resolved_title`,
     instead of feed :attr:`~Feed.user_title` or :attr:`~Feed.title`.
+  * Allow filtering entries by the entry source.
 
 * Fix :meth:`~Reader.enable_search` / :meth:`~Reader.update_search`
   not working when the search database is missing but change tracking is enabled